requirement.txt
crewai
crewai_tools
load_dotenv
langchain-huggingface
youtube-transcript-api==0.6.1
yt_dlp==2023.11.14